pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "claw-forge"
version = "0.1.0a1"
description = "Multi-provider autonomous coding agent harness with API rotation pool"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [{ name = "ClawInfra", email = "alex.chen31337@gmail.com" }]
keywords = ["ai", "agent", "coding", "claude", "llm", "multi-provider"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Code Generators",
]
dependencies = [
    "typer>=0.12,<1",
    "rich>=13,<14",
    "httpx>=0.27,<1",
    "pydantic>=2,<3",
    "pydantic-settings>=2,<3",
    "pyyaml>=6,<7",
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.30,<1",
    "sqlalchemy[asyncio]>=2,<3",
    "aiosqlite>=0.20,<1",
    "anthropic>=0.40,<1",
    "openai>=1.50,<2",
    "sse-starlette>=2,<3",
]

[project.optional-dependencies]
bedrock = ["boto3>=1.35,<2", "anthropic[bedrock]>=0.40,<1"]
azure = ["azure-identity>=1.17,<2", "openai>=1.50,<2"]
vertex = ["anthropic[vertex]>=0.40,<1", "google-auth>=2,<3"]
all-providers = ["claw-forge[bedrock,azure,vertex]"]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.24,<1",
    "pytest-cov>=5,<6",
    "ruff>=0.6,<1",
    "mypy>=1.11,<2",
]

[project.scripts]
claw-forge = "claw_forge.cli:app"

[project.entry-points."claw_forge.plugins"]
initializer = "claw_forge.plugins.initializer:InitializerPlugin"
coding = "claw_forge.plugins.coding:CodingPlugin"
testing = "claw_forge.plugins.testing:TestingPlugin"
reviewer = "claw_forge.plugins.reviewer:ReviewerPlugin"

[project.urls]
Homepage = "https://github.com/clawinfra/claw-forge"
Repository = "https://github.com/clawinfra/claw-forge"
Issues = "https://github.com/clawinfra/claw-forge/issues"

[tool.hatch.build.targets.wheel]
packages = ["claw_forge"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.coverage.run]
source = ["claw_forge"]
branch = true

[tool.coverage.report]
fail_under = 90
